Runs as intended, but minor formating and readability updates are needed

+++ MidTermProj(Scoopa)/Player.cpp
@@ -97,7 +97,7 @@ public:
 	}
 
 	int getNumberOfCards() {
-		return earnedCards.size();
+		return static_cast<int>(earnedCards.size());
 	}
 
 	int getGold() {
